group tracking stuff

+++ zone/client.h
@@ -964,7 +964,7 @@ public:
 
 	// Configurable Tracking Skill
 	uint16 GetClassTrackingDistanceMultiplier(uint16 class_);
-	float  GetTrackingDistance();
+	int  GetTrackingDistance();
 
 	bool CanThisClassTrack();
